Streamline code to conform shellcheck and various improvements

+++ stage2/00-copies-and-fills/02-run.sh
@@ -1,6 +1,3 @@
 #!/bin/bash -e
 
-if [ -f "${ROOTFS_DIR}/etc/ld.so.preload" ]; then
-   mv "${ROOTFS_DIR}/etc/ld.so.preload" "${ROOTFS_DIR}/etc/ld.so.preload.disabled"
-fi
-
+[ -f "${ROOTFS_DIR}/etc/ld.so.preload" ] && mv "${ROOTFS_DIR}/etc/ld.so.preload" "${ROOTFS_DIR}/etc/ld.so.preload.disabled"
